Moving all code in Jabberwocky.Glass.Autofac.Mvc that does not have an explicit dependency on Autofac to Jabberwocky.Glass.Mvc.
