+++ src/Jabberwocky.Glass.Autofac.Mvc/Extensions/MvcRegistrationExtensions.cs
@@ -5,8 +5,8 @@ using Autofac.Extras.AttributeMetadata;
 using Autofac.Integration.Mvc;
 using Glass.Mapper.Sc;
 using Glass.Mapper.Sc.ModelCache;
-using Jabberwocky.Glass.Autofac.Mvc.Models;
 using Jabberwocky.Glass.Autofac.Mvc.Models.Factory;
+using Jabberwocky.Glass.Mvc.Models;
 using Jabberwocky.Glass.Mvc.Services;
 using Jabberwocky.Glass.Mvc.Util;
 using IViewModelFactory = Jabberwocky.Glass.Mvc.Models.Factory.IViewModelFactory;
