+++ src/Jabberwocky.Glass.Mvc/Jabberwocky.Glass.Mvc.csproj
@@ -58,6 +58,9 @@
     </Reference>
     <Reference Include="System" />
     <Reference Include="System.Core" />
+    <Reference Include="System.Reflection, Version=4.1.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a, processorArchitecture=MSIL">
+      <HintPath>..\..\packages\System.Reflection.4.1.0\lib\net462\System.Reflection.dll</HintPath>
+    </Reference>
     <Reference Include="System.Web" />
     <Reference Include="System.Web.Helpers, Version=3.0.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35, processorArchitecture=MSIL">
       <HintPath>..\..\packages\Microsoft.AspNet.WebPages.3.2.3\lib\net45\System.Web.Helpers.dll</HintPath>
@@ -94,7 +97,14 @@
     <Compile Include="Attributes\HttpGetOrInvalidHttpPostAttribute.cs" />
     <Compile Include="Attributes\ValidateFormHandlerAttribute.cs" />
     <Compile Include="Attributes\ValidHttpPostAttribute.cs" />
+    <Compile Include="Extensions\HtmlHelperExtensions.cs" />
+    <Compile Include="Models\Attributes\AllowNestedDatasourceAttribute.cs" />
+    <Compile Include="Models\Attributes\ConfigureDatasourceAttribute.cs" />
+    <Compile Include="Models\Attributes\DisableNestedDatasourceAttribute.cs" />
+    <Compile Include="Models\Factory\ViewModelFactory.cs" />
     <Compile Include="Models\Factory\IViewModelFactory.cs" />
+    <Compile Include="Models\GlassViewModel.cs" />
+    <Compile Include="Models\InjectableGlassViewModelBase.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Services\DatasourceNestingOptions.cs" />
     <Compile Include="Services\IRenderingContextService.cs" />
@@ -103,6 +113,7 @@
     <Compile Include="Views\CustomGlassView.cs" />
   </ItemGroup>
   <ItemGroup>
+    <None Include="app.config" />
     <None Include="packages.config" />
   </ItemGroup>
   <ItemGroup>
