+++ src/Jabberwocky.Glass.Mvc/Models/GlassViewModel.cs
@@ -1,4 +1,4 @@
-﻿namespace Jabberwocky.Glass.Autofac.Mvc.Models
+﻿namespace Jabberwocky.Glass.Mvc.Models
 {
 	public abstract class GlassViewModel<TGlassModel> : InjectableGlassViewModelBase where TGlassModel : class
 	{
