+++ src/Jabberwocky.Glass.Mvc/Properties/AssemblyInfo.cs
@@ -35,3 +35,5 @@ using System.Runtime.InteropServices;
 [assembly: AssemblyVersion("3.0.0.0")]
 [assembly: AssemblyFileVersion("3.0.0.0")]
 [assembly: AssemblyInformationalVersion("3.0.0-beta")]
+
+[assembly: InternalsVisibleTo("Jabberwocky.Glass.Mvc.Tests")]
