+++ src/Jabberwocky.Glass.Mvc/packages.config
@@ -8,4 +8,11 @@
   <package id="Microsoft.Web.Infrastructure" version="1.0.0.0" targetFramework="net452" />
   <package id="Sitecore.Kernel.NoReferences" version="9.0.171002" targetFramework="net462" developmentDependency="true" />
   <package id="Sitecore.Mvc.NoReferences" version="9.0.171002" targetFramework="net462" developmentDependency="true" />
+  <package id="System.ComponentModel" version="4.0.1" targetFramework="net462" />
+  <package id="System.Diagnostics.Debug" version="4.0.11" targetFramework="net462" />
+  <package id="System.Globalization" version="4.0.11" targetFramework="net462" />
+  <package id="System.Linq" version="4.1.0" targetFramework="net462" />
+  <package id="System.Linq.Expressions" version="4.1.0" targetFramework="net462" />
+  <package id="System.Reflection" version="4.1.0" targetFramework="net462" />
+  <package id="System.Resources.ResourceManager" version="4.0.1" targetFramework="net462" />
 </packages>
